pyproject.toml
[project]
name = "ai-agent-intensive-course-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.4.1",
    "google-api-core>=2.25.1",
    "google-genai>=1.21.0",
]
